tests for postitanalysis
