+++ CountPostIts.ImageRecognition.Tests/CountPostIts.ImageRecognition.Tests.csproj
@@ -77,10 +77,10 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="InformationFeatureTest.cs" />
-    <Compile Include="InformationTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ColourRangesTest.cs" />
     <Compile Include="ImageFilterTest.cs" />
+    <Compile Include="PostItAnalysisTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
